# Conflicts:
#	eth-withdrawal/src/main/kotlin/withdrawalservice/WithdrawalServiceImpl.kt
#	eth/src/main/kotlin/notary/endpoint/eth/EthRefundStrategyImpl.kt
#	notary-btc-integration-test/src/main/kotlin/integration/helper/BtcIntegrationHelperUtil.kt
#	notary-commons/src/main/kotlin/registration/IrohaAccountCreator.kt
#	notary-commons/src/main/kotlin/registration/RegistrationServiceEndpoint.kt
#	notary-commons/src/main/kotlin/sidechain/iroha/util/ModelUtil.kt
#	notary-commons/src/main/kotlin/sidechain/iroha/util/RequestHelper.kt
#	notary-commons/src/test/java/registration/IrohaAccountCreatorTest.java
#	notary-eth-integration-test/src/main/kotlin/integration/helper/EthIntegrationHelperUtil.kt
#	notary-sora-integration-test/src/integration-test/kotlin/integration/sora/SoraIntegrationTest.kt

+++ notary-eth-integration-test/src/integration-test/kotlin/integration/eth/DepositMultiIntegrationTest.kt
@@ -2,13 +2,9 @@ package integration.eth
 
 import config.IrohaCredentialConfig
 import config.loadEthPasswords
-
 import integration.helper.EthIntegrationHelperUtil
 import integration.helper.IrohaConfigHelper
-import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.Assertions
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
+import org.junit.jupiter.api.*
 import provider.eth.ETH_PRECISION
 import sidechain.iroha.CLIENT_DOMAIN
 import sidechain.iroha.util.ModelUtil
@@ -103,7 +99,7 @@ class DepositMultiIntegrationTest {
             integrationHelper.waitOneIrohaBlock()
 
             Assertions.assertEquals(
-                BigDecimal(amount, ETH_PRECISION.toInt()).add(BigDecimal(initialAmount)),
+                BigDecimal(amount, ETH_PRECISION).add(BigDecimal(initialAmount)),
                 BigDecimal(integrationHelper.getIrohaAccountBalance(clientIrohaAccountId, etherAssetId))
             )
         }
